requirements.txt
Flask==2.0.3
requests==2.26.0
pydantic==1.9.0
google-auth==2.3.3
google-api-python-client==2.31.0